admin

+++ server/src/main/java/com/server/studio27/controllers/StudentController.java
@@ -34,7 +34,7 @@ public class StudentController {
 
             for (Map<String, Object> row : rows) {
                 Map<String, Object> studentMap = new HashMap<>();
-                studentMap.put("studentId", ((Integer) row.get("studentId")).intValue());
+                studentMap.put("studentId", ((Long) row.get("studentId")).intValue());
                 studentMap.put("email", (String) row.get("email"));
                 studentMap.put("password", (String) row.get("password"));
                 studentMap.put("ime", (String) row.get("ime"));
@@ -48,7 +48,7 @@ public class StudentController {
                                     join pohadja p on p.kursId = k.kursId
                                     where studentId = ?
                                     """;
-                List<Map<String, Object>> kursRows = jdbcTemplate.queryForList(SQL2, ((Integer) row.get("studentId")).intValue());
+                List<Map<String, Object>> kursRows = jdbcTemplate.queryForList(SQL2, ((Long) row.get("studentId")).intValue());
                 List<Map<String, Object>> kursevi = new ArrayList<>();
                 for (Map<String, Object> kursRow : kursRows) {
                     Map<String, Object> kursMap = new HashMap<>();
